control_panel/requirements.txt
flask
flask-cors
flask-WTF
Flask-Session
boto3
paho-mqtt
kafka-python